Extracted irk utils
