+++ esphome_nimble/components/nimble_tracker/nimble_tracker.h
@@ -1,8 +1,6 @@
 #include "esphome/core/component.h"
 #include "queue.h"
 #include "NimBLEDevice.h"
-#include "NimBLEScan.h"
-#include "NimBLEAdvertisedDevice.h"
 
 namespace esphome
 {
